+++ crates/cairo-profiler/tests/e2e.rs
@@ -1,6 +1,7 @@
 use assert_fs::fixture::PathCopy;
 use indoc::indoc;
 use snapbox::cmd::{cargo_bin, Command as SnapboxCommand};
+use std::str;
 use test_case::test_case;
 
 #[test]
@@ -79,5 +80,152 @@ fn missing_syscall_from_versioned_constants_file() {
                 Missing libfuncs cost in versioned constants file: [CallContract].
                 Make sure to include costs of these libfuncs in the aforementioned file.
             "
+        ));
+}
+
+#[test]
+fn view_samples() {
+    let project_root = project_root::get_project_root().unwrap();
+    let temp_dir = assert_fs::TempDir::new().unwrap();
+    temp_dir
+        .copy_from(
+            project_root.join("crates/cairo-profiler/tests/contracts/balance_simple/precompiled/"),
+            &["*.json"],
+        )
+        .unwrap();
+
+    SnapboxCommand::new(cargo_bin!("cairo-profiler"))
+        .current_dir(&temp_dir)
+        .arg("build-profile")
+        .arg("trace_balance_simple.json")
+        .assert()
+        .success();
+
+    let output = SnapboxCommand::new(cargo_bin!("cairo-profiler"))
+        .current_dir(&temp_dir)
+        .arg("view")
+        .arg("profile.pb.gz")
+        .arg("--list-samples")
+        .assert()
+        .success()
+        .get_output()
+        .stdout
+        .clone();
+    let output_str = str::from_utf8(&output).expect("Output was not valid utf-8");
+
+    assert!(
+        output_str.contains("steps"),
+        "Output contains: {}, missing 'steps'",
+        &output_str
+    );
+    assert!(
+        output_str.contains("calls"),
+        "Output contains: {}, missing 'calls'",
+        &output_str
+    );
+    assert!(
+        output_str.contains("memory holes"),
+        "Output contains: {}, missing 'memory holes'",
+        &output_str
+    );
+    assert!(
+        output_str.contains("range check builtin"),
+        "Output contains: {}, missing 'range check builtin'",
+        &output_str
+    );
+}
+
+#[test]
+fn view_steps() {
+    let project_root = project_root::get_project_root().unwrap();
+    let temp_dir = assert_fs::TempDir::new().unwrap();
+    temp_dir
+        .copy_from(
+            project_root.join("crates/cairo-profiler/tests/contracts/balance_simple/precompiled/"),
+            &["*.json"],
+        )
+        .unwrap();
+
+    SnapboxCommand::new(cargo_bin!("cairo-profiler"))
+        .current_dir(&temp_dir)
+        .arg("build-profile")
+        .arg("trace_balance_simple.json")
+        .assert()
+        .success();
+
+    // stdout asserts were generated using `go tool pprof -top profile.pb.gz` command
+    // when changing any view_* tests please always generate expected output using this tool
+    // formatting was changed manually, since it differs a bit between pprof and cairo-profiler view
+
+    SnapboxCommand::new(cargo_bin!("cairo-profiler"))
+        .current_dir(&temp_dir)
+        .arg("view")
+        .arg("profile.pb.gz")
+        .arg("--limit")
+        .arg("2137")
+        .arg("--sample")
+        .arg("steps")
+        .assert()
+        .success()
+        .stdout_eq(indoc!(
+            r#"
+            
+            Showing nodes accounting for 1371 steps, 100.00% of 1371 steps total
+            Showing top 14 nodes out of 14
+            
+                  flat |  flat% |    sum% |        cum |    cum% |  
+            -----------+--------+---------+------------+---------+-----------------------------------------------------------------------------------------------
+             827 steps | 60.32% |  60.32% |  827 steps |  60.32% | "CallContract" 
+              91 steps |  6.64% |  66.96% |  168 steps |  12.25% | "core::result::ResultSerde::::deserialize" 
+              87 steps |  6.35% |  73.30% |   87 steps |   6.35% | "StorageRead" 
+              75 steps |  5.47% |  78.77% |   75 steps |   5.47% | "snforge_std::_cheatcode::handle_cheatcode" 
+              53 steps |  3.87% |  82.64% | 1246 steps |  90.88% | "balance_simple_integrationtest::test_contract::test_cannot_increase_balance_with_zero_value" 
+              39 steps |  2.84% |  85.49% |   39 steps |   2.84% | "core::array::SpanFelt252Serde::deserialize" 
+              38 steps |  2.77% |  88.26% |   38 steps |   2.77% | "snforge_std::cheatcodes::contract_class::DeclareResultSerde::deserialize" 
+              36 steps |  2.63% |  90.88% |  144 steps |  10.50% | "snforge_std::cheatcodes::contract_class::declare" 
+              34 steps |  2.48% |  93.36% |  170 steps |  12.40% | "snforge_std::cheatcodes::contract_class::ContractClassImpl::deploy" 
+              33 steps |  2.41% |  95.77% |  120 steps |   8.75% | "balance_simple::HelloStarknet::__wrapper__HelloStarknetImpl__get_balance" 
+              27 steps |  1.97% |  97.74% |  341 steps |  24.87% | "balance_simple_integrationtest::test_contract::deploy_contract" 
+              26 steps |  1.90% |  99.64% |   26 steps |   1.90% | "core::array::serialize_array_helper::" 
+               5 steps |  0.36% | 100.00% | 1371 steps | 100.00% | "Contract: SNFORGE_TEST_CODE\nFunction: SNFORGE_TEST_CODE_FUNCTION\n" 
+               0 steps |  0.00% | 100.00% |  120 steps |   8.75% | "Contract: HelloStarknet\nFunction: get_balance\n" 
+            "#
+        ));
+}
+
+#[test]
+fn view_range_check_builtin() {
+    let project_root = project_root::get_project_root().unwrap();
+    let temp_dir = assert_fs::TempDir::new().unwrap();
+    temp_dir
+        .copy_from(
+            project_root.join("crates/cairo-profiler/tests/contracts/balance_simple/precompiled/"),
+            &["*.json"],
+        )
+        .unwrap();
+
+    SnapboxCommand::new(cargo_bin!("cairo-profiler"))
+        .current_dir(&temp_dir)
+        .arg("build-profile")
+        .arg("trace_balance_simple.json")
+        .arg("--view")
+        .arg("--sample")
+        .arg("range check builtin")
+        .arg("--limit")
+        .arg("3")
+        .assert()
+        .success()
+        .stdout_eq(indoc!(
+            r#"
+            
+            Showing nodes accounting for 36 range check builtin, 97.30% of 37 range check builtin total
+            Showing top 3 nodes out of 14
+            
+                               flat |  flat% |   sum% |                    cum |    cum% |  
+            ------------------------+--------+--------+------------------------+---------+-----------------------------------------------------------------------
+             19 range check builtin | 51.35% | 51.35% | 37 range check builtin | 100.00% | "Contract: SNFORGE_TEST_CODE\nFunction: SNFORGE_TEST_CODE_FUNCTION\n" 
+             15 range check builtin | 40.54% | 91.89% | 15 range check builtin |  40.54% | "CallContract" 
+              2 range check builtin |  5.41% | 97.30% |  3 range check builtin |   8.11% | "Contract: HelloStarknet\nFunction: get_balance\n" 
+            "#
         ));
 }